+++ tests/src/historical_tests.cpp
@@ -10,6 +10,7 @@
 #include <cstdint>
 #include <cstdlib>
 #include <filesystem>
+#include <optional>
 #include <stdexcept>  // logic_error
 #include <utility>    // move
 
@@ -240,8 +241,8 @@ TEST_F(HistoricalTests, TestBatchDownloadAll) {
   const TempFile temp_dbn_file{tmp_path_ / "job123/test.dbn"};
   mock_server_.MockGetJson("/v0/batch.list_files", {{"job_id", kJobId}},
                            kListFilesResp);
-  mock_server_.MockStreamDbn("/v0/job_id/test.dbn", {},
-                             TEST_DATA_DIR "/test_data.mbo.v3.dbn");
+  mock_server_.MockGetDbn("/v0/job_id/test.dbn", {},
+                          TEST_DATA_DIR "/test_data.mbo.v3.dbn");
   mock_server_.MockGetJson("/v0/job_id/test_metadata.json", {{"key", "value"}});
   const auto port = mock_server_.ListenOnThread();
 
@@ -424,7 +425,7 @@ TEST_F(HistoricalTests, TestMetadataGetDatasetCondition) {
                               {"last_modified_date", "2023-03-01"}},
                              {{"date", "2022-11-10"},
                               {"condition", "missing"},
-                              {"last_modified_date", "2023-03-01"}}};
+                              {"last_modified_date", nullptr}}};
   mock_server_.MockGetJson("/v0/metadata.get_dataset_condition",
                            {{"dataset", dataset::kXnasItch},
                             {"start_date", "2022-11-06"},
@@ -440,7 +441,7 @@ TEST_F(HistoricalTests, TestMetadataGetDatasetCondition) {
       {"2022-11-07", DatasetCondition::Available, "2023-03-01"},
       {"2022-11-08", DatasetCondition::Degraded, "2023-03-01"},
       {"2022-11-09", DatasetCondition::Pending, "2023-03-01"},
-      {"2022-11-10", DatasetCondition::Missing, "2023-03-01"},
+      {"2022-11-10", DatasetCondition::Missing, std::nullopt},
   };
   EXPECT_EQ(res, kExp);
 }
@@ -464,8 +465,22 @@ TEST_F(HistoricalTests, TestMetadataListUnitPrices) {
 }
 
 TEST_F(HistoricalTests, TestMetadataGetDatasetRange) {
-  const nlohmann::json kResp = {{"start", "2017-05-21T00:00:00.000000000Z"},
-                                {"end", "2022-12-01T00:00:00.000000000Z"}};
+  const nlohmann::json kResp = {
+      {"start", "2017-05-21T00:00:00.000000000Z"},
+      {"end", "2022-12-01T00:00:00.000000000Z"},
+      {"schema",
+       {
+           {"bbo-1m",
+            {{"start", "2020-08-02T00:00:00.000000000Z"},
+             {"end", "2023-03-23T00:00:00.000000000Z"}}},
+           {"ohlcv-1s",
+            {{"start", "2020-08-02T00:00:00.000000000Z"},
+             {"end", "2023-03-23T00:00:00.000000000Z"}}},
+           {"ohlcv-1m",
+            {{"start", "2020-08-02T00:00:00.000000000Z"},
+             {"end", "2023-03-23T00:00:00.000000000Z"}}},
+
+       }}};
   mock_server_.MockGetJson("/v0/metadata.get_dataset_range",
                            {{"dataset", dataset::kXnasItch}}, kResp);
   const auto port = mock_server_.ListenOnThread();
@@ -625,17 +640,17 @@ TEST_F(HistoricalTests, TestSymbologyResolve) {
 }
 
 TEST_F(HistoricalTests, TestTimeseriesGetRange_Basic) {
-  mock_server_.MockStreamDbn("/v0/timeseries.get_range",
-                             {{"dataset", dataset::kGlbxMdp3},
-                              {"symbols", "ESH1"},
-                              {"schema", "mbo"},
-                              {"start", "1609160400000711344"},
-                              {"end", "1609160800000711344"},
-                              {"encoding", "dbn"},
-                              {"stype_in", "raw_symbol"},
-                              {"stype_out", "instrument_id"},
-                              {"limit", "2"}},
-                             TEST_DATA_DIR "/test_data.mbo.v3.dbn.zst");
+  mock_server_.MockPostDbn("/v0/timeseries.get_range",
+                           {{"dataset", dataset::kGlbxMdp3},
+                            {"symbols", "ESH1"},
+                            {"schema", "mbo"},
+                            {"start", "1609160400000711344"},
+                            {"end", "1609160800000711344"},
+                            {"encoding", "dbn"},
+                            {"stype_in", "raw_symbol"},
+                            {"stype_out", "instrument_id"},
+                            {"limit", "2"}},
+                           TEST_DATA_DIR "/test_data.mbo.v3.dbn.zst");
   const auto port = mock_server_.ListenOnThread();
 
   databento::Historical target{&logger_, kApiKey, "localhost",
@@ -661,16 +676,16 @@ TEST_F(HistoricalTests, TestTimeseriesGetRange_Basic) {
 }
 
 TEST_F(HistoricalTests, TestTimeseriesGetRange_NoMetadataCallback) {
-  mock_server_.MockStreamDbn("/v0/timeseries.get_range",
-                             {{"dataset", dataset::kGlbxMdp3},
-                              {"start", "2022-10-21T13:30"},
-                              {"end", "2022-10-21T20:00"},
-                              {"symbols", "CYZ2"},
-                              {"schema", "tbbo"},
-                              {"encoding", "dbn"},
-                              {"stype_in", "raw_symbol"},
-                              {"stype_out", "instrument_id"}},
-                             TEST_DATA_DIR "/test_data.tbbo.v3.dbn.zst");
+  mock_server_.MockPostDbn("/v0/timeseries.get_range",
+                           {{"dataset", dataset::kGlbxMdp3},
+                            {"start", "2022-10-21T13:30"},
+                            {"end", "2022-10-21T20:00"},
+                            {"symbols", "CYZ2"},
+                            {"schema", "tbbo"},
+                            {"encoding", "dbn"},
+                            {"stype_in", "raw_symbol"},
+                            {"stype_out", "instrument_id"}},
+                           TEST_DATA_DIR "/test_data.tbbo.v3.dbn.zst");
   const auto port = mock_server_.ListenOnThread();
 
   databento::Historical target{&logger_, kApiKey, "localhost",
@@ -689,7 +704,7 @@ TEST_F(HistoricalTests, TestTimeseriesGetRange_NoMetadataCallback) {
 TEST_F(HistoricalTests, TestTimeseriesGetRange_BadRequest) {
   const nlohmann::json resp{
       {"detail", "Authorization failed: illegal chars in username."}};
-  mock_server_.MockBadRequest("/v0/timeseries.get_range", resp);
+  mock_server_.MockBadPostRequest("/v0/timeseries.get_range", resp);
   const auto port = mock_server_.ListenOnThread();
 
   databento::Historical target{&logger_, kApiKey, "localhost",
@@ -714,8 +729,8 @@ TEST_F(HistoricalTests, TestTimeseriesGetRange_BadRequest) {
 }
 
 TEST_F(HistoricalTests, TestTimeseriesGetRange_CallbackException) {
-  mock_server_.MockStreamDbn("/v0/timeseries.get_range", {},
-                             TEST_DATA_DIR "/test_data.mbo.v3.dbn.zst");
+  mock_server_.MockPostDbn("/v0/timeseries.get_range", {},
+                           TEST_DATA_DIR "/test_data.mbo.v3.dbn.zst");
   const auto port = mock_server_.ListenOnThread();
 
   databento::Historical target{&logger_, kApiKey, "localhost",
@@ -732,8 +747,8 @@ TEST_F(HistoricalTests, TestTimeseriesGetRange_CallbackException) {
 }
 
 TEST_F(HistoricalTests, TestTimeseriesGetRange_Cancellation) {
-  mock_server_.MockStreamDbn("/v0/timeseries.get_range", {},
-                             TEST_DATA_DIR "/test_data.mbo.v3.dbn.zst");
+  mock_server_.MockPostDbn("/v0/timeseries.get_range", {},
+                           TEST_DATA_DIR "/test_data.mbo.v3.dbn.zst");
   const auto port = mock_server_.ListenOnThread();
 
   databento::Historical target{&logger_, kApiKey, "localhost",
@@ -762,9 +777,9 @@ TEST_F(HistoricalTests, TestTimeseriesGetRange_LargeChunks) {
                    10005,
                    {}}};
   constexpr auto kRecordCount = 50'000;
-  mock_server_.MockStreamDbn("/v0/timeseries.get_range",
-                             {{"dataset", ToString(Dataset::IfusImpact)}},
-                             Record{&mbp1.hd}, kRecordCount, 75'000);
+  mock_server_.MockPostDbn("/v0/timeseries.get_range",
+                           {{"dataset", ToString(Dataset::IfusImpact)}},
+                           Record{&mbp1.hd}, kRecordCount, 75'000);
   const auto port = mock_server_.ListenOnThread();
 
   databento::Historical target{&logger_, kApiKey, "localhost",
@@ -789,9 +804,9 @@ TEST_F(HistoricalTests, TestTimeseriesGetRange_UnreadBytes) {
                    10005,
                    {}}};
   constexpr auto kRecordCount = 1'000;
-  mock_server_.MockStreamDbn("/v0/timeseries.get_range",
-                             {{"dataset", ToString(Dataset::IfusImpact)}},
-                             Record{&mbp1.hd}, kRecordCount, 20, 75'000);
+  mock_server_.MockPostDbn("/v0/timeseries.get_range",
+                           {{"dataset", ToString(Dataset::IfusImpact)}},
+                           Record{&mbp1.hd}, kRecordCount, 20, 75'000);
   const auto port = mock_server_.ListenOnThread();
 
   logger_ = mock::MockLogReceiver{[](auto count, LogLevel level,
@@ -815,16 +830,16 @@ TEST_F(HistoricalTests, TestTimeseriesGetRange_UnreadBytes) {
 }
 
 TEST_F(HistoricalTests, TestTimeseriesGetRangeToFile) {
-  mock_server_.MockStreamDbn("/v0/timeseries.get_range",
-                             {{"dataset", dataset::kGlbxMdp3},
-                              {"start", "2022-10-21T13:30"},
-                              {"end", "2022-10-21T20:00"},
-                              {"symbols", "CYZ2"},
-                              {"schema", "tbbo"},
-                              {"encoding", "dbn"},
-                              {"stype_in", "raw_symbol"},
-                              {"stype_out", "instrument_id"}},
-                             TEST_DATA_DIR "/test_data.tbbo.v3.dbn.zst");
+  mock_server_.MockPostDbn("/v0/timeseries.get_range",
+                           {{"dataset", dataset::kGlbxMdp3},
+                            {"start", "2022-10-21T13:30"},
+                            {"end", "2022-10-21T20:00"},
+                            {"symbols", "CYZ2"},
+                            {"schema", "tbbo"},
+                            {"encoding", "dbn"},
+                            {"stype_in", "raw_symbol"},
+                            {"stype_out", "instrument_id"}},
+                           TEST_DATA_DIR "/test_data.tbbo.v3.dbn.zst");
   const auto port = mock_server_.ListenOnThread();
 
   databento::Historical target{&logger_, kApiKey, "localhost",